Moved "Format.hxx" include file to ${CMAKE_INSTALL_PREFIX}/include/SoDa/Format.hxx.
It used to be in "${CMAKE...}/include/SoDaFormat/Format.hxx

This is part of an effort to move all SoDa related includes (SoDaFormat, SoDaSignals,
SoDaRadio) into a common include directory.

Also fixed a minor formatting issue with floating point.  This code still needs to be improved wrt "rounding" and truncation.